pubsub-gcp/requirements.txt
google-cloud-pubsub<1